Update Proto.TimePicker.js
By pressing down key, time option moves down in the list but scroll bar does not moves down. Due to this, selected time option is not visible.

Steps to Reproduce:
1. Click on the time picker text box.
5. Scroll down using the down key.

ACTUAL RESULT:

The time selected moves down the list but the list does not scroll down to show the selected option on pressing the down key.

EXPECTED RESULT:

The time selected should moves down the list and list should scroll down to show the selected option on pressing the down key.